Fix for Android IL2CPP builds.
Any P/Invoke methods need to be linkable on IL2CPP builds for android.
It seems facebook is already aware of this by already moving out the
P/Invoke methods to a separate dll. However, facebook was not utilizing
this.

I did not add the GetIOSWrapper function, only appended to it. It was
not in use originally. I replaced the default constructor IOSFacebook to
use GetIOSWrapper instead of the redundantly packed IOSWrapper.cs

IOSWrapper.cs was removed from the project as well. I believe this was
Facebooks intention originally.