Fixes crash at startup.
I am unaware of why reflection was neccisary here, I didn't know it was
neccisary to hide usage of UnityEngine.AndroidJavaClass (even on non
android platforms) but the use of reflection was assumidly breaking
IL2CPP builds which could not know Facebooks intention of using them at
link time. The app crashes at the call to GetGenericArguments in
CallStaticMethodInfoGeneric get accessor for what its worth.

I've tested this fix, and it does not crash anymore.