Trash Dash Changes
Summary:
This adds a number of changed added in June / July, including testing network access by making a GraphAPI call to the /me endpoint to obtain the player's name and adding Interstitial Ads and Rewarded Videos to the store. See IAPHandler.cs

Most other updated files are artifacts of updating the Unity SDK version and/or building the APK.

Already uploaded the new APK to the portal.

Differential Revision: D22541136

fbshipit-source-id: 00885a7b977bd576bdf95a8318a2e821fe26aa58