Unity SDK Windows Catalog Implementation
Summary:
We started to implement existing methods also in the implementation for android mobile.
Since from now on they will be common, we take the base definition of the interface and the class, in this case, to obtain the catalog.
We have verified that the compilation works correctly and that the getcatalog method appears in the Unity project.

Reviewed By: carloslopezmdez

Differential Revision: D30930792

fbshipit-source-id: 7973f86c0b0b1cfbeee8908e08494cf59de631d5