Fix win_configure script for Unity 2019 and above on Windows
Summary:
Because we are having issues Unity DLLs on Windows platform. We decided to force unity version to work with UnitySDK project on Windows.

So, only Unity 2017.4.40f1 can be used to work with UnitySDK project on Windows devices.

This is only for add UnitySDK features using Windows Devices for WIndows Platform.

For other platform you must to wok on a MAC.

Reviewed By: ffgomezfb

Differential Revision: D37279438

fbshipit-source-id: 299d199a9c12a34c770365e5a2d074ce0c758e41