Fixed issue with Windows devices and build targets
Summary:
Added a checker when the user launch the app in editor mode.

This check will check the build target platform and set an internal variable with this information.

This internal variable is used to launch Windows SDK only if  is running on windows devices and build target is windows.

Reviewed By: SergioGuerreroFB

Differential Revision: D37859022

fbshipit-source-id: 4b549cbc44e0f5e6dc499fc96efb5c360317137b