Fixed InvalidCastException when parsing IAP prices
Summary:
Based on a bug report in the Unity SDK repo: https://github.com/facebook/facebook-sdk-for-unity/issues/656

Sometimes an IAP product can have a round number, without decimals, causing this line to throw the exception:
`InvalidCastException: Unable to cast object of type 'Int64' to type 'Double'`

For example, 20.99$ is translated to 1250₽ (russian rubles), which triggers this exception. We can avoid this exception by adding Convert.ToDouble, as suggested by the developer.

Differential Revision: D41649757

fbshipit-source-id: 12b28fa31c802ca867da3509a67ce608c8034c0c