Reorganising external dependency manager
Summary:
Reorganising  external dependency manager to make easy to NOT import to a current project with and existing plugin.

NOTE: the motivation of this fix came from the task T143437219. Was too difficult to remove our current version of External Dependency Manager when a dev has a conflict with other SDK like FIrebase.

Reviewed By: SergioGuerreroFB

Differential Revision: D43453556

fbshipit-source-id: b0e66cefdd7db6859ff1b9cc9eb4e322aefd0749