Now disregarding already set AndroidSdkRoot if SdkUseEmbedded is enabled
In fresh Unity installations the `AndroidSdkRoot` editor pref is not set. They rather default to `SdkUseEmbedded` set to true.

Before the fix, we were unable to build unless an SDK path was explicitly set, even if the "Use Embedded SDK" option was enabled. 

Facebook SDK was not allowing to export a project without an Android SDK Root path explicitly set.