Lay IAP Foundation for Unity Test App
Summary: We need to test IAP in FBSDK Unity. In order to do this, we will create a new screen where the IAP events will be triggered from. Then, we can open this screen and trigger some events. The screen will contain the ability to purchase a consumable, non-consumable, and subscription product. This screen is created in this diff.

Reviewed By: jjiang10

Differential Revision: D72600217

fbshipit-source-id: dfe22ef524ab7a4cbb610c71a8de722c786a81d6