Build IAP iOS Unity Plugin and test IAP
Summary: In order to test IAP for unity, we need to build an iOS native plugin that calls the iOS IAP APIs. This diff writes the Unity Wrapper that will enable this native plugin to be used in the test iOS app. This diff also fully implements the IAP functionality in the test app.

Reviewed By: jjiang10

Differential Revision: D72600216

fbshipit-source-id: d79f011c9bbcbb77e5ae6895bbc33a1b96978c98