Make IAP Wrapper More Generic
Summary: The IAP testing functionality in the test app needs to support testing multiple integrations (i.e. SK1, SK2, and Android). In this diff, we make the IAP wrapper that wraps the native implementations more generic so that we can support all of these integrations. We also update the IAP test screen to support initializing SK1, SK2, and GPBL, depending on the device. The next series of diffs will actually natively implement SK2 and GPBL so we can test all scenarios.

Reviewed By: jjiang10

Differential Revision: D72609468

fbshipit-source-id: a1abb5aaa7b228354e6ecb8d9bc69225bda45d72