Add package for multiple targets
Summary:
Meta:
Our manifest merger is outdated, dating back to 2015.
To improve its functionality, we need to decouple it from AOSP and upgrade to the latest version.
Here's an initial attempt at doing so: D83352021
However, with the new manifest mergers, every library's AndroidManifest.xml should have a package field, which many of our repository's libraries currently lack.
We must address this issue before we can successfully land the diff which is what this stack of diff is about.
More: T239793438

Reviewed By: hick209

Differential Revision: D86414311

fbshipit-source-id: 2e3e58ba3531ee6c05cfebd735306e45ccc189c0